Rename runner files: remove _oo suffix.

+++ template.py
@@ -15,11 +15,11 @@ for day in range(1, 26):
 
     for part in range(1, 3):
         solution_path = folder_path + f"/day_{day}_part_{part}.cpp"
-        # if os.path.isfile(solution_path):
-        #     continue
+        if os.path.isfile(solution_path):
+            continue
 
         with open(solution_path, 'w', encoding="UTF-8") as solution:
-            solution.write(f"""#include "../runner_oo.hpp"
+            solution.write(f"""#include "../runner.hpp"
 #include "../utils.hpp"
 #include <string>
 #include <optional>
